// src/models/User.ts
export type Token = {
  token: string;
};
export type UserProfileToken = {
  name: string;
  token: string;
};

export type UserProfile = {
  username: string;
};

export type userLogin = {
  username: string;
  password: string;
};

export type userRegister = {
  name: string;
  email: string;
  password: string;
};